Test classification pipeline
